refactor: move components and make more generic to enable use between download and share
