+++ src/components/DownloadComponents/ResourceCard/ResourceCard.test.tsx
@@ -3,12 +3,12 @@ import userEvent from "@testing-library/user-event";
 
 import renderWithTheme from "../../../__tests__/__helpers__/renderWithTheme";
 
-import DLCard from "./DLCard";
+import ResourceCard from "./ResourceCard";
 
-describe("DownloadCard", () => {
+describe("ResourceCard", () => {
   it("renders a checkbox with an icon, title and resource file type", () => {
     renderWithTheme(
-      <DLCard
+      <ResourceCard
         id="unique-123"
         name="downloadResources"
         label="Worksheet"
@@ -34,7 +34,7 @@ describe("DownloadCard", () => {
     };
 
     const { rerender } = renderWithTheme(
-      <DLCard
+      <ResourceCard
         id="unique-123"
         checked={value}
         name="downloadResources"
@@ -54,7 +54,7 @@ describe("DownloadCard", () => {
     await user.click(label);
 
     rerender(
-      <DLCard
+      <ResourceCard
         id="unique-123"
         checked={value}
         name="downloadResources"
