+++ src/components/DownloadComponents/ResourcePageLayout/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ResourcePageLayout";
